Optimized move functionality for animals!
