--- src/main/java/Erik/animals/predators/Wolf.java
@@ -1,13 +1,0 @@
-package Erik.animals.predators;
-
-import Erik.animals.Predator;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-public class Wolf extends Predator {
-
-    public Wolf(){
-
-    }
-
-
-}
