+++ src/main/java/erik/animals/Herbivore.java
@@ -1,4 +1,4 @@
-package Erik.animals;
+package erik.animals;
 
 public abstract class Herbivore extends Animal {
     public Herbivore(){
