+++ src/main/java/erik/animals/Plant.java
@@ -1,4 +1,4 @@
-package Erik.animals;
+package erik.animals;
 
 public class Plant extends Animal{
