+++ src/main/java/erik/animals/predators/Fox.java
@@ -1,6 +1,6 @@
-package Erik.animals.predators;
+package erik.animals.predators;
 
-import Erik.animals.Predator;
+import erik.animals.Predator;
 
 public class Fox extends Predator {
     public Fox(){
