+++ src/main/java/erik/animals/predators/Snake.java
@@ -1,6 +1,6 @@
-package Erik.animals.predators;
+package erik.animals.predators;
 
-import Erik.animals.Predator;
+import erik.animals.Predator;
 
 public class Snake extends Predator {
     public Snake(){
