+++ src/main/java/erik/animals/predators/Wolf.java
@@ -1,0 +1,12 @@
+package erik.animals.predators;
+
+import erik.animals.Predator;
+
+public class Wolf extends Predator {
+
+    public Wolf(){
+
+    }
+
+
+}
